Update Board.java
Added an IF statement to the calcAdjacencies method to handle the situation wherein a Player object was forcibly moved into a room by another Player's guess. This logic will allow the Player to stay in the room to make a guess. Also changed the movePlayer method to set the forceMove variable in the Player object to indicate the player was forcibly moved